Refactoring, add ResponseDto
