+++ Spring/src/main/java/GFO/Spring/domain/User/presentation/dto/response/SignupResponseDto.java
@@ -1,0 +1,12 @@
+package GFO.Spring.domain.User.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class SignupResponseDto {
+    private String email;
+    private String name;
+    private String password;
+    private String duty;
+    private int classNum;
+}
