--- Spring/src/main/java/GFO/Spring/domain/student/repository/StudentRepo.java
@@ -1,8 +1,0 @@
-package GFO.Spring.domain.student.repository;
-
-import GFO.Spring.domain.student.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepo extends JpaRepository<Student, String> {
-
-}
